test integer query

+++ app/Http/Controllers/HasilStudiController.php
@@ -31,21 +31,21 @@ class HasilStudiController extends Controller
     public function detail($username)
     {
         $tahun_ajaran = TahunAjaran::latest()->get();
-        $mahasiswa = User::where('username', (int)$username)->firstOrFail();
+        $mahasiswa = User::where('username', $username)->firstOrFail();
         return view('hasil-studi.detail', compact('mahasiswa', 'tahun_ajaran'));
     }
 
     public function edit($username, $khs_id)
     {
-        $mahasiswa = User::where('username', (int)$username)->firstOrFail();
-        $khs = KartuStudi::where('username', (int)$username)->where('id', $khs_id)->firstOrFail();
+        $mahasiswa = User::where('username', $username)->firstOrFail();
+        $khs = KartuStudi::where('username', $username)->where('id', $khs_id)->firstOrFail();
         return view('hasil-studi.edit', compact('mahasiswa', 'khs'));
     }
 
     public function update(Request $request, $username, $khs_id)
     {
-        $mahasiswa = User::where('username', (int)$username)->firstOrFail();
-        $khs = KartuStudi::where('username', (int)$username)->where('id', $khs_id)->firstOrFail();
+        $mahasiswa = User::where('username', $username)->firstOrFail();
+        $khs = KartuStudi::where('username', $username)->where('id', $khs_id)->firstOrFail();
         $mata_kuliah = MataKuliah::where('kode', $khs->kode_mata_kuliah)->first();
         $request->validate([
             'tugas' => 'numeric|max:100',
@@ -57,38 +57,38 @@ class HasilStudiController extends Controller
         $huruf = null;
         $bobot = null;
         if ($request->tugas && $request->uts && $request->uas && $mata_kuliah) {
-            $angka = ((int)$request->tugas * 30 +  (int)$request->uts * 30 + (int)$request->uas * 40) / 100;
+            $angka = ((float)$request->tugas * 30 +  (float)$request->uts * 30 + (float)$request->uas * 40) / 100;
             $angka = number_format($angka, 2);
             $huruf = 'E';
-            $bobot = 0 * (int)$mata_kuliah->jumlah_sks;
+            $bobot = 0 * (float)$mata_kuliah->jumlah_sks;
 
             if ($angka >= 85) {
                 $huruf = "A";
-                $bobot = 4 * (int)$mata_kuliah->jumlah_sks;
+                $bobot = 4 * (float)$mata_kuliah->jumlah_sks;
             } else if ($angka >= 80) {
                 $huruf = 'A-';
-                $bobot = 3.75 * (int)$mata_kuliah->jumlah_sks;
+                $bobot = 3.75 * (float)$mata_kuliah->jumlah_sks;
             } else if ($angka >= 75) {
                 $huruf = 'B+';
-                $bobot = 3.5 * (int)$mata_kuliah->jumlah_sks;
+                $bobot = 3.5 * (float)$mata_kuliah->jumlah_sks;
             } else if ($angka >= 70) {
                 $huruf = 'B';
-                $bobot = 3 * (int)$mata_kuliah->jumlah_sks;
+                $bobot = 3 * (float)$mata_kuliah->jumlah_sks;
             } else if ($angka >= 65) {
                 $huruf = 'B-';
-                $bobot = 2.70 * (int)$mata_kuliah->jumlah_sks;
+                $bobot = 2.70 * (float)$mata_kuliah->jumlah_sks;
             } else if ($angka >= 60) {
                 $huruf = 'C+';
-                $bobot = 2.35 * (int)$mata_kuliah->jumlah_sks;
+                $bobot = 2.35 * (float)$mata_kuliah->jumlah_sks;
             } else if ($angka >= 55) {
                 $huruf = 'C';
-                $bobot = 2 * (int)$mata_kuliah->jumlah_sks;
+                $bobot = 2 * (float)$mata_kuliah->jumlah_sks;
             } else if ($angka >= 50) {
                 $huruf = 'D+';
-                $bobot = 1.50 * (int)$mata_kuliah->jumlah_sks;
+                $bobot = 1.50 * (float)$mata_kuliah->jumlah_sks;
             } else if ($angka >= 40) {
                 $huruf = 'D';
-                $bobot = 1 * (int)$mata_kuliah->jumlah_sks;
+                $bobot = 1 * (float)$mata_kuliah->jumlah_sks;
             }
         }
 
@@ -107,7 +107,7 @@ class HasilStudiController extends Controller
 
     public function api_get_khs($username, $tahun_ajaran)
     {
-        $khs = KartuStudi::where('username', (int)$username)->where('tahun_ajaran', $tahun_ajaran)->with('mata_kuliah')->latest()->get();
+        $khs = KartuStudi::where('username', $username)->where('tahun_ajaran', $tahun_ajaran)->with('mata_kuliah')->latest()->get();
 
         $ip = 0;
         $total_bobot = 0;
